feat: 💥 add `faas upload` command, refactor `faas deploy` and other faas command params
- added `faas upload` for independent uploads.
- refactor `faas deploy`, now equals `faas build` + `faas upload`.
- refactor `faas` command params for better clarity.
