+++ packages/cli/src/commands/faas/debug.ts
@@ -1,15 +1,6 @@
-import fs from 'fs';
-import path from 'path';
-import chalk from 'chalk';
-import { AxiosError } from 'axios';
-
-import { readFile } from 'fs/promises';
 import type { Command } from 'commander';
-import { program as globalProgram } from 'commander';
 
-import { createLogger, resolveCwdAbsolutePath } from '@/utils';
-import { createCloudFunction, updateCloudFunction } from '@/request/api';
-import { buildFunction } from './build';
+import { createLogger } from '@/utils';
 import { deployFunction } from './deploy';
 import { invokeMockData } from './mock';
 
@@ -19,7 +10,7 @@ const logger = createLogger(COMMAND_NAME, 'verbose');
 // TODO: There are variables here that can be shared.
 const defaultConfig = {
   entryDir: './src/function',
-  distDir: './dist',
+  builtDir: './dist',
   outputDir: './log',
   mockDir: './mock',
 };
@@ -42,13 +33,13 @@ export function registerCommand(program: Command) {
       defaultConfig.entryDir,
     )
     .option(
-      '--dist-dir <value>',
+      '--built-dir <value>',
       'Cloud function built file output directory',
-      defaultConfig.distDir,
+      defaultConfig.builtDir,
     )
     .option(
       '--output-dir <value>',
-      '云函数构建文件输出目录',
+      'log output directory',
       defaultConfig.outputDir,
     )
     .option('--mock-dir <value>', 'mock data directory', defaultConfig.mockDir)
@@ -61,14 +52,15 @@ export function registerCommand(program: Command) {
         return;
       }
 
-      const { entryDir, distDir, outputDir, mockDir } = options;
-      await buildFunction({
-        entryDir,
-        outputDir: distDir,
+      const { entryDir, builtDir, outputDir, mockDir } = options;
+      // Deploy the selected or provided cloud function with specified directories for entry and build.
+      await deployFunction({
         functionName: name,
+        entryDir,
+        builtDir,
       });
 
-      await deployFunction({ deployDir: distDir, functionName: name });
+      // Invoke the cloud function with mock data, specifying the mock data directory and the output directory for logs.
       await invokeMockData({
         functionName: name,
         dir: mockDir,
