+++ packages/cli/src/commands/faas/index.ts
@@ -1,6 +1,7 @@
 import type { Command } from 'commander';
 
 import { registerCommand as registerBuildCommand } from './build';
+import { registerCommand as registerUploadCommand } from './upload';
 import { registerCommand as registerDeployCommand } from './deploy';
 import { registerCommand as registerMockCommand } from './mock';
 import { registerCommand as registerDebugCommand } from './debug';
@@ -15,6 +16,7 @@ export function registerCommand(program: Command) {
   // 注册子命令 (或许可以尝试从自动化导入的方式来简化代码)
   const tasks = [
     registerBuildCommand,
+    registerUploadCommand,
     registerDeployCommand,
     registerMockCommand,
     registerDebugCommand,
