+++ packages/cli/src/commands/faas/mock.ts
@@ -71,17 +71,26 @@ export async function invokeMockData({
   functionName,
   ...options
 }: BuildFaasParams) {
-  const folderPath = resolveCwdAbsolutePath(options.dir);
+  if (!functionName) {
+    const message = 'Please input cloud function name!';
+    logger.error(COMMAND_NAME, message);
 
+    return Promise.reject(new Error(message));
+  }
+
+  const folderPath = resolveCwdAbsolutePath(options.dir);
   const mockPath = path.resolve(folderPath, `${functionName}.js`);
+
   logger.verbose(COMMAND_NAME, 'mockPath:', mockPath);
 
   let data = {};
 
   // check mock file exists
   if (!fs.existsSync(mockPath)) {
+    logger.warn(COMMAND_NAME, `${mockPath} mock file not found.`);
     logger.warn(
-      `${chalk.bold(functionName)} mock file not found. \n`,
+      COMMAND_NAME,
+      `${mockPath} mock file not found.`,
       'This call will use default value `{}` to invoke cloud function.',
     );
   } else {
@@ -115,9 +124,9 @@ export async function invokeMockData({
       dirPath: dir,
       content,
     });
-    logger.info(
+    logger.success(
       COMMAND_NAME,
-      'invoke result log saved to:',
+      'Invoke result log saved to:',
       path.join(dir, fileName),
     );
 
@@ -139,7 +148,7 @@ export async function invokeMockData({
       if (err) throw err;
       logger.error(
         COMMAND_NAME,
-        'Invoke cloud function error.\n',
+        'Invoke cloud function error.',
         `Log saved to: ${errorPath}`,
       );
     });
@@ -147,21 +156,25 @@ export async function invokeMockData({
 }
 
 /**
- * invoke cloud function with mock data
+ * Invoke cloud function with mock data
  *
  * @example
- * invoke cloud function
+ * Invoke the "createUser" cloud function without specifying a mock directory:
  * ```
  * mincloudx faas mock createUser
  * ```
  *
  * @example
- * If the default mock folder is not used and wish to change folder
- * you can use `--dir` option change:
+ * Invoke the "createUser" cloud function with mock data from a specific directory:
+ * ```
+ * mincloudx faas mock createUser --dir ./mocks
+ * ```
  *
- *  ```
- *  mincloudx faas mock createUser --dir ./mocks
- *  ```
+ * @example
+ * Invoke the "createUser" cloud function and specify a directory for the output logs:
+ * ```
+ * mincloudx faas mock createUser --out ./logs
+ * ```
  */
 export function registerCommand(program: Command) {
   return program
